Code refactored for mantainability
