+++ index.html
@@ -105,7 +105,7 @@
     </li>
   </template>
   <footer>Made by Aldehil Sánchez</footer>
-  <script src="./js/main.js" type="module"></script>
+  <script src="./main.js" type="module"></script>
 </body>
 
 </html>
