+++ pages/login.html
@@ -35,8 +35,7 @@
     </form>
   </main>
   <footer>Made by Aldehil Sánchez</footer>
-  <script src="../js/label-top-animation.js" type="module"></script>
-  <script src="../js/login.js" type="module"></script>
+  <script src="../js/scripts/login.js" type="module"></script>
 </body>
 
 </html>
