+++ pages/sign-in.html
@@ -42,8 +42,7 @@
     </form>
   </main>
   <footer>Made by Aldehil Sánchez</footer>
-  <script src="../js/label-top-animation.js" type="module"></script>
-  <script src="../js/signIn.js" type="module"></script>
+  <script src="../js/scripts/signIn.js" type="module"></script>
 </body>
 
 </html>
